+++ .gitignore
@@ -27,3 +27,4 @@ sarasas5-kietiakiai.txt
 sarasas5-vargsiukai.txt
 sarasas5.txt
 .gitignore
+v1_0-main.exe
